# docs/openapi_inference.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CarMate 视觉推理服务",
    "description": "车载智能视觉系统后端 - 交警手势识别 (ctpgr-pytorch)",
    "version": "2.0.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health Check",
        "operationId": "health_check_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/police-gesture/stream/reset": {
      "post": {
        "summary": "Reset Police Gesture Stream",
        "operationId": "reset_police_gesture_stream_api_police_gesture_stream_reset_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_reset_police_gesture_stream_api_police_gesture_stream_reset_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/police-gesture/stream/frame": {
      "post": {
        "summary": "Recognize Police Gesture Stream Frame",
        "operationId": "recognize_police_gesture_stream_frame_api_police_gesture_stream_frame_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_recognize_police_gesture_stream_frame_api_police_gesture_stream_frame_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/police-gesture/preview": {
      "post": {
        "summary": "Create Police Gesture Preview",
        "description": "将上传视频转为浏览器友好的 MP4(H.264/yuv420p)，仅用于前端预览播放。",
        "operationId": "create_police_gesture_preview_api_police_gesture_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_police_gesture_preview_api_police_gesture_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/police-gesture/recognize": {
      "post": {
        "summary": "Recognize Police Gesture",
        "description": "交警手势识别 - 上传图片或视频",
        "operationId": "recognize_police_gesture_api_police_gesture_recognize_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_recognize_police_gesture_api_police_gesture_recognize_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/police-gesture/recognize/stream": {
      "post": {
        "summary": "Recognize Police Gesture Stream Video",
        "description": "流式视频识别: 边分析边返回采样帧结果，最后返回平滑后的完整时间线。",
        "operationId": "recognize_police_gesture_stream_video_api_police_gesture_recognize_stream_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_recognize_police_gesture_stream_video_api_police_gesture_recognize_stream_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plate/recognize": {
      "post": {
        "summary": "Recognize Plate",
        "operationId": "recognize_plate_api_plate_recognize_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_recognize_plate_api_plate_recognize_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/driver-gesture/recognize": {
      "post": {
        "summary": "Recognize Driver Gesture",
        "operationId": "recognize_driver_gesture_api_driver_gesture_recognize_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_recognize_driver_gesture_api_driver_gesture_recognize_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_create_police_gesture_preview_api_police_gesture_preview_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_police_gesture_preview_api_police_gesture_preview_post"
      },
      "Body_recognize_driver_gesture_api_driver_gesture_recognize_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_recognize_driver_gesture_api_driver_gesture_recognize_post"
      },
      "Body_recognize_plate_api_plate_recognize_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_recognize_plate_api_plate_recognize_post"
      },
      "Body_recognize_police_gesture_api_police_gesture_recognize_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_recognize_police_gesture_api_police_gesture_recognize_post"
      },
      "Body_recognize_police_gesture_stream_frame_api_police_gesture_stream_frame_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "stream_id": {
            "type": "string",
            "title": "Stream Id",
            "default": "default"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_recognize_police_gesture_stream_frame_api_police_gesture_stream_frame_post"
      },
      "Body_recognize_police_gesture_stream_video_api_police_gesture_recognize_stream_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_recognize_police_gesture_stream_video_api_police_gesture_recognize_stream_post"
      },
      "Body_reset_police_gesture_stream_api_police_gesture_stream_reset_post": {
        "properties": {
          "stream_id": {
            "type": "string",
            "title": "Stream Id",
            "default": "default"
          }
        },
        "type": "object",
        "title": "Body_reset_police_gesture_stream_api_police_gesture_stream_reset_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}